# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API documentation for Project.",
        "title": "Qpay API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/commission": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all commissions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all commissions",
                "responses": {
                    "200": {
                        "description": "List of commissions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.CommissionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in getting commissions\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new commission with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a new commission",
                "parameters": [
                    {
                        "description": "Commission details",
                        "name": "commissionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CommissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created commission details",
                        "schema": {
                            "$ref": "#/definitions/handlers.CommissionResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Bind Error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in creating commission\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/admin/commission/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific commission by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get commission by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Commission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commission details",
                        "schema": {
                            "$ref": "#/definitions/handlers.CommissionResponse"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"Commission not found!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/admin/gateway": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all gateways.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all gateways",
                "responses": {
                    "200": {
                        "description": "List of gateways",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.GatewayResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in getting gateways\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/admin/gateway/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific gateway by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get gateway by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gateway details",
                        "schema": {
                            "$ref": "#/definitions/handlers.GatewayResponse"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"Gateway not found!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the status of a specific gateway by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update gateway status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated status details",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.StatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated gateway details",
                        "schema": {
                            "$ref": "#/definitions/handlers.GatewayResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Status field is unsupported\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/admin/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register a new admin with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a new admin",
                "parameters": [
                    {
                        "description": "Admin details",
                        "name": "adminRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AdminRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success message",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Bind Error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "403": {
                        "description": "{\"status\": \"error\", \"message\": \"Invalid admin details\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "409": {
                        "description": "{\"status\": \"error\", \"message\": \"Username or email already exists\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in create admin\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/admin/transaction": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all transactions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all transactions",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.TransactionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in getting transactions\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
        "/admin/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all users.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.UserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in getting users\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AdminRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "handlers.BankAccountResponse": {
            "type": "object",
            "properties": {
                "account_owner": {
                    "type": "string"
                },
                "bank_logo": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "sheba": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.CommissionRequest": {
            "type": "object",
            "properties": {
                "amount_per_transaction": {
                    "type": "number"
                },
                "percent_per_transaction": {
                    "type": "number"
                }
            }
        },
        "handlers.CommissionResponse": {
            "type": "object",
            "properties": {
                "Percent_per_transaction": {
                    "type": "number"
                },
                "amount_per_transaction": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.GatewayResponse": {
            "type": "object",
            "properties": {
                "bank_account": {
                    "$ref": "#/definitions/handlers.BankAccountResponse"
                },
                "commission": {
                    "$ref": "#/definitions/handlers.CommissionResponse"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "route": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.StatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.TransactionResponse": {
            "type": "object",
            "properties": {
                "payment_amount": {
                    "type": "number"
                },
                "payment_date": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "purchaser_card": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tracking_code": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UserResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "identity": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}